# api/openapi.yaml
openapi: 3.1.0
info: {title: Stock God Local Web API, version: 2.0.0, description: 'Local stock prediction, retained market data,
    settings and audit APIs.'}
servers:
- {url: 'http://127.0.0.1:34115'}
paths:
...
  /api/v1/system/shutdown:
    post:
      operationId: shutdownSystem
      responses:
        '200':
          description: Shutdown requested
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommandResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/settings:
    get:
      operationId: getSettings
      responses:
        '200': {description: Settings}
    put:
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/JsonObject'}
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommandResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/prediction/ai/configs:
    get:
      operationId: listPredictionAIConfigs
      responses:
        '200': {description: Enabled models belonging to the requested center}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/prediction/settings:
    get:
      operationId: getPredictionSettings
      responses:
        '200':
          description: Independent research configuration
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PredictionSettings'}
        '400': {$ref: '#/components/responses/BadRequest'}
    put:
      operationId: updatePredictionSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PredictionSettings'}
      responses:
        '200':
          description: Saved configuration and new revision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PredictionSettings'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/prediction/ai/configs/test:
    post:
      operationId: testPredictionAIConfig
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AIConfigTestRequest'}
      responses:
        '200': {description: AI configuration test result}
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/prediction/analysis-runs/{id}/audit:
    get:
      operationId: getPredictionAnalysisRunAudit
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Immutable redacted audit detail or an explicit legacy-unavailable result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PredictionAuditDetail'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
  /api/v1/prediction/analysis-runs/{id}/audit/export:
    get:
      operationId: exportPredictionAnalysisRunAudit
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Redacted immutable audit ZIP bundle
          content:
            application/zip:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
  /api/v1/prediction/replays:
    post:
      operationId: createPredictionReplay
      description: Replays one immutable evidence and prompt snapshot with another model configuration without writing
        formal recommendations, trades, positions, or accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreatePredictionReplayRequest'}
      responses:
        '202':
          description: Replay accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PredictionReplay'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalServerError'}
  /api/v1/prediction/replays/{id}:
    get:
      operationId: getPredictionReplay
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Replay state
          immutable result: null
          and comparison summary: null
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PredictionReplay'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
components:
  responses:
    BadRequest:
      description: Invalid JSON request or parameters
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Conflict:
      description: Request conflicts with the current resource revision
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    InternalServerError:
      description: Request failed because of an internal service or storage error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Forbidden:
      description: Request is not from the bound loopback origin
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
    PredictionSettings:
      type: object
      additionalProperties: false
      required: [revision, config, aiConfigs]
      properties:
        revision: {type: integer, minimum: 1}
        config: {$ref: '#/components/schemas/JsonObject'}
        aiConfigs:
          type: array
          items: {$ref: '#/components/schemas/JsonObject'}
    JsonObject: {type: object, description: Provider- or configuration-defined JSON object whose stable fields are
        not owned by this transport layer.}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: {type: string}
    CommandResponse:
      type: object
      required: [ok]
      properties:
        ok: {type: boolean}
        message: {type: string}
    AIConfigTestRequest:
      type: object
      required: [id]
      properties:
        id: {type: integer}
...
    PredictionAuditRunState:
      type: object
      required: [status, payloadCount, lastError, createdAt, updatedAt]
      properties:
        status:
          type: string
          enum: [capturing, complete, failed, legacy_unavailable]
        payloadCount: {type: integer, minimum: 0}
        lastError: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    PredictionAuditPayload:
      type: object
      required: [payloadId, phase, callSequence, attempt, providerName, modelName, modelParameters, finalPrompt,
        finalPromptSha256, evidenceSnapshot, evidenceSha256, tools, rawResponse, rawResponseSha256, repairedResponse,
        repairedResponseSha256, repairLog, redactionCount, createdAt]
      properties:
        payloadId: {type: string}
        phase: {type: string}
        callSequence: {type: integer, minimum: 1}
        attempt: {type: integer, minimum: 1}
        providerName: {type: string}
        modelName: {type: string}
        modelParameters: {$ref: '#/components/schemas/JsonObject'}
        finalPrompt: {type: string}
        finalPromptSha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        evidenceSnapshot: {type: string}
        evidenceSha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        tools:
          type: array
          items: {type: string}
        rawResponse: {type: string}
        rawResponseSha256: {type: string, pattern: '^$|^[0-9a-f]{64}$'}
        repairedResponse: {type: string}
        repairedResponseSha256: {type: string, pattern: '^$|^[0-9a-f]{64}$'}
        repairLog: {type: string}
        redactionCount: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
    PredictionAuditDetail:
      type: object
      required: [availability, ownerType, ownerId, state, payloads]
      properties:
        availability:
          type: string
          enum: [available, legacy_unavailable]
        ownerType:
          type: string
          enum: [research2, replay]
        ownerId: {type: string}
        cutoffAt:
          oneOf:
          - {type: string, format: date-time}
          - {type: 'null'}
        state: {$ref: '#/components/schemas/PredictionAuditRunState'}
        payloads:
          type: array
          items: {$ref: '#/components/schemas/PredictionAuditPayload'}
    CreatePredictionReplayRequest:
      type: object
      required: [sourceOwnerId, modelConfigId]
      properties:
        sourceOwnerId: {type: string}
        modelConfigId: {type: integer, minimum: 1}
    PredictionReplay:
      type: object
      required: [replayId, sourceOwnerType, sourceOwnerId, modelConfigId, status, cutoffAt, createdAt, lastError,
        result, resultSha256, diffSummary]
      properties:
        replayId: {type: string}
        sourceOwnerType:
          type: string
          enum: [research2]
        sourceOwnerId: {type: string}
        modelConfigId: {type: integer, minimum: 1}
        status:
          type: string
          enum: [queued, running, completed, failed]
        cutoffAt: {type: string, format: date-time}
        createdAt: {type: string, format: date-time}
        startedAt:
          oneOf:
          - {type: string, format: date-time}
          - {type: 'null'}
        completedAt:
          oneOf:
          - {type: string, format: date-time}
          - {type: 'null'}
        lastError: {type: string}
        result: {type: string}
        resultSha256: {type: string, pattern: '^$|^[0-9a-f]{64}$'}
        diffSummary: {$ref: '#/components/schemas/JsonObject'}
...